backend/app/requirements.txt
aiofiles==24.1.0
azure-storage-blob==12.23.1
azure-identity==1.19.0
azure-ai-formrecognizer==3.3.3
azure-search-documents==11.5.2
azure-core==1.34.0
azure-cosmos
nest_asyncio==1.6.0
psutil==7.0.0
PyPDF2==3.0.1
openai==1.75.0
crewai==0.121.0
crewai-tools==0.45.0
python-dotenv==1.0.1

imagehash==4.3.1
opencv-python==4.10.0.84
torch==2.5.1
torchvision==0.20.1
open-clip-torch==2.32.0
scikit-learn==1.5.2
requests